MISUV-***:   * - ClaimToAdjustService~ refactoring;
