+++ app/controllers/claimToAdjustPoa/CheckYourAnswersController.scala
@@ -25,8 +25,8 @@ import models.claimToAdjustPoa.{PaymentOnAccountViewModel, PoAAmendmentData, Sel
 import models.core.{CheckMode, Nino}
 import play.api.Logger
 import play.api.mvc._
-import services.claimToAdjustPoa.RecalculatePoaHelper
-import services.{ClaimToAdjustPoaCalculationService, ClaimToAdjustService, PaymentOnAccountSessionService}
+import services.claimToAdjustPoa.{ClaimToAdjustPoaCalculationService, RecalculatePoaHelper}
+import services.{ClaimToAdjustService, PaymentOnAccountSessionService}
 import uk.gov.hmrc.auth.core.AuthorisedFunctions
 import utils.{AuthenticatorPredicate, ClaimToAdjustUtils}
 import views.html.claimToAdjustPoa.CheckYourAnswers
