+++ app/services/claimToAdjustPoa/ClaimToAdjustHelper.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package services
+package services.claimToAdjustPoa
 
 import auth.MtdItUser
 import connectors.{CalculationListConnector, ChargeHistoryConnector}
@@ -26,12 +26,11 @@ import models.core.Nino
 import models.financialDetails.DocumentDetail
 import models.incomeSourceDetails.TaxYear
 import models.incomeSourceDetails.TaxYear.makeTaxYearWithEndYear
+import services.DateServiceInterface
 import uk.gov.hmrc.http.{HeaderCarrier, InternalServerException}
 
 import java.time.{LocalDate, Month}
-import scala.::
 import scala.concurrent.{ExecutionContext, Future}
-import scala.util.Try
 
 // TODO: This part of the logic expected to be moved within BE
 // TODO: plain models like: TaxYear and PaymentOnAccountViewModel will be return via new connector
